Crate iroh_base

Base types and utilities for Iroh

Modules

base32base32

ticketticket

Structs

Hashhash

Hash type used throughout.

HashAndFormathash

A hash and format pair

NodeAddrkey

Network-level addressing information for an iroh node.

PublicKeykey

A public key.

RelayMaprelay

Configuration of all the relay servers that can be used.

RelayNoderelay

Information on a specific relay server.

RelayQuicConfigrelay

Configuration for speaking to the QUIC endpoint on the relay server to do QUIC address discovery.

RelayUrlrelay

A URL identifying a relay server.

SecretKeykey

A secret key.

SharedSecretkey

Shared Secret.

Signaturekey

Ed25519 signature.

Enums

BlobFormathash

A format identifier

KeyParsingErrorkey

Error when deserialising a PublicKey or a SecretKey.

Constants

DEFAULT_RELAY_QUIC_PORTrelay

The default QUIC port used by the Relay server to accept QUIC connections for QUIC address discovery

DEFAULT_STUN_PORTrelay

The default STUN port used by the Relay server.

PUBLIC_KEY_LENGTHkey

The length of an ed25519 PublicKey, in bytes.

Type Aliases

NodeIdkey

The identifier for a node in the (iroh) network.