Module iroh_base::key

Cryptographic key handling for iroh-net.

Modules

encryption 🔒

The private and public keys of a node.

Structs

CryptoKeys 🔒

PublicKey

A public key.

SecretKey

A secret key.

SharedSecret

Shared Secret.

Signature

Ed25519 signature.

Enums

KeyParsingError

Error when deserialising a PublicKey or a SecretKey.

Constants

KEY_CACHE_CAPACITY 🔒

Maximum number of keys in the crypto key cache. CryptoKeys are 224 bytes, keys are 32 bytes, so each entry is 256 bytes plus some overhead.

KEY_CACHE_TTL 🔒

Expiry time for the crypto key cache.

PUBLIC_KEY_LENGTH

The length of an ed25519 PublicKey, in bytes.

Statics

KEY_CACHE 🔒

Functions

get_or_create_crypto_keys 🔒

Get or create the crypto keys, and project something out of them.

lock_key_cache 🔒

Type Aliases

NodeId

The identifier for a node in the (iroh) network.