Struct iroh_base::key::Signature

#[repr(C)]
pub struct Signature { /* private fields */ }

Available on crate feature key only.

Ed25519 signature.

This type represents a container for the byte serialization of an Ed25519 signature, and does not necessarily represent well-formed field or curve elements.

Signature verification libraries are expected to reject invalid field elements at the time a signature is verified.

Implementations

impl Signature

pub const BYTE_SIZE: usize = 64usize

Size of an encoded Ed25519 signature in bytes.

pub fn from_bytes(bytes: &[u8; 64]) -> Signature

Parse an Ed25519 signature from a byte slice.

pub fn from_components(R: [u8; 32], s: [u8; 32]) -> Signature

Parse an Ed25519 signature from its R and s components.

pub fn from_slice(bytes: &[u8]) -> Result<Signature, Error>

Parse an Ed25519 signature from a byte slice.

Returns

Ok on success

Err if the input byte slice is not 64-bytes

pub fn r_bytes(&self) -> &[u8; 32]

Bytes for the R component of a signature.

pub fn s_bytes(&self) -> &[u8; 32]

Bytes for the s component of a signature.

pub fn to_bytes(&self) -> [u8; 64]

Return the inner byte array.

pub fn to_vec(&self) -> Vec<u8>

Available on crate feature alloc only.

Convert this signature into a byte vector.