iroh_base
Module relay_map

Available on crate feature relay only.

based on tailscale/tailcfg/derpmap.go

Structs

QuicConfig

Configuration for speaking to the QUIC endpoint on the relay server to do QUIC address discovery.

RelayMap

Configuration of all the relay servers that can be used.

RelayNode

Information on a specific relay server.

RelayUrl

A URL identifying a relay server.

Constants

DEFAULT_RELAY_QUIC_PORT

The default QUIC port used by the Relay server to accept QUIC connections for QUIC address discovery

DEFAULT_STUN_PORT

The default STUN port used by the Relay server.