iroh_base::ticket
Struct NodeTicket

pub struct NodeTicket { /* private fields */ }

Available on crate feature ticket only.

A token containing information for establishing a connection to a node.

Contains

The NodeId of the node to connect to (a 32-byte ed25519 public key).

If used, the [โRelayUrl`] of on which the node can be reached.

Any direct addresses on which the node might be reachable.

This allows establishing a connection to the node in most circumstances where it is possible to do so.

This NodeTicket is a single item which can be easily serialized and deserialized and implements the Ticket trait. The Display and FromStr traits can also be used to round-trip the ticket to string.

Implementations

impl NodeTicket

pub fn new(node: NodeAddr) -> Self

Creates a new ticket.

pub fn node_addr(&self) -> &NodeAddr

The NodeAddr of the provider for this ticket.