Crate iroh_base

Base types and utilities for Iroh

Modules

ticketticket

Tickets is a serializable object combining information required for an operation. Typically tickets contain all information required for an operation, e.g. an iroh blob ticket would contain the hash of the data as well as information about how to reach the provider.

Structs

NodeAddrkey

Network-level addressing information for an iroh node.

PublicKeykey

A public key.

RelayUrlrelay

A URL identifying a relay server.

RelayUrlParseErrorrelay

Can occur when parsing a string into a RelayUrl.

SecretKeykey

A secret key.

Signaturekey

Ed25519 signature.

Enums

KeyParsingErrorkey

Error when deserialising a PublicKey or a SecretKey.

Type Aliases

NodeIdkey

The identifier for a node in the (iroh) network.

SignatureErrorkey

Errors which may occur while processing signatures and keypairs.