iroh_base::ticket

Trait Ticket

pub trait Ticket: Sized {
    const KIND: &'static str;

    // Required methods
    fn to_bytes(&self) -> Vec<u8> ;
    fn from_bytes(bytes: &[u8]) -> Result<Self, ParseError>;

    // Provided methods
    fn serialize(&self) -> String { ... }
    fn deserialize(str: &str) -> Result<Self, ParseError> { ... }
}

Available on crate feature ticket only.

A ticket is a serializable object combining information required for an operation.

Tickets support serialization to a string using base32 encoding. The kind of ticket will be prepended to the string to make it somewhat self describing.

Versioning is left to the implementer. Some kinds of tickets might need versioning, others might not.

The serialization format for converting the ticket from and to bytes is left to the implementer. We recommend using postcard for serialization.

Required Associated Constants

const KIND: &'static str

String prefix describing the kind of iroh ticket.

This should be lower case ascii characters.

Required Methods

fn to_bytes(&self) -> Vec<u8>

Serialize to bytes used in the base32 string representation.

fn from_bytes(bytes: &[u8]) -> Result<Self, ParseError>

Deserialize from the base32 string representation bytes.

Provided Methods

fn serialize(&self) -> String

Serialize to string.

fn deserialize(str: &str) -> Result<Self, ParseError>

Deserialize from a string.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.